gigacommit with data structures, beginning of a debug fct. and more
(I know gigacommits are not exactly coolio but this is a private dumb little project at the moment...)

+++ src/util.cpp
@@ -1,6 +1,6 @@
 #include <zephyr/kernel.h>
 
-void hexdump(const uint8_t *buf, uint32_t len) {
+void hexdump(const uint8_t *buf, size_t len) {
     for (int i = 0; i < len; i++) {
         printf("%02x", buf[i]);
         if (i < len - 1) {
@@ -9,7 +9,7 @@ void hexdump(const uint8_t *buf, uint32_t len) {
     }
 }
 
-void hexdump(const char *prefix, const uint8_t *buf, uint32_t len) {
+void hexdump(const char *prefix, const uint8_t *buf, size_t len) {
     fputs(prefix, stdout);
     hexdump(buf, len);
 }
